+++ src/estd/internal/string/operators.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../fwd/string.h"
+
+namespace estd {
+
+}